Prevent empty book creation, limit num. of pages
Added a condition to the "Create book" button event listener, that
checks if Title, Author, and Number of Pages inputs are empty. If any of
this are empty, an alert pops up and reminds the user to finish filling
the form or close it.

Added a condition to the createBook prototype function that replaces the
text displayed if the number of pages is bigger than:
999999999999999999999999999999 but the value inside the object remains
unchanged.